python/openmlconverter/pyproject.toml
[project]
name = "mlcommons_openmlconverter"
description = "Converting OpenML dataset schema to the new MLCommons format."
version = "0.1"
authors = [
    { name = "Jos van der Velde", email = "j.d.v.d.velde@tue.nl"},
]
dependencies = [
    "langcodes",
    "language_data",
    "python-dateutil"
]
readme = "README.md"

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
    "responses",
    "types-requests",
    "types-python-dateutil"
]

[tool.black]
line-length = 100